// src/helpers/constants/type.ts
export type RequestBodyType<D> = { param?: string | number; data?: D; success?: (data?: any) => void; error?: (data?: any) => void };